MNT: move server to top-level service module
