+++ databroker/service/metadata.py
@@ -5,7 +5,7 @@ import ujson
 import json
 import types
 
-from ....headersource.core import NoRunStop, NoEventDescriptors, NoRunStart
+from ..headersource.core import NoRunStop, NoEventDescriptors, NoRunStart
 
 loop = tornado.ioloop.IOLoop.instance()
